Tested the new_redis_client.rs and used it to replace the redis.rs file.  We'll develop this version going forward.
