+++ apps/framework-cli/src/infrastructure/redis.rs
@@ -1,2 +1,1 @@
-pub mod new_redis_client;
 pub mod redis_client;
